Move to IndexedSetoid.
Using elem-of, and the generic IInverse (for IndexedInverse; should
probably be expanded and then made into an infix operator), BagEq'
can be easily defined.  It is a properly "representative independent"
version of BagEq.

Hopefully this won't bring back subst juggling.  Nicely, most of the
commented out lemmas about Some will not actually be needed.  Their
BagEq' versions will be quite straightforward to do.

The biggest 'to do' is going to be to make ParComp deal with
IndexedSetoid.