Add lots and lots of comments.
Build parallel composition in two ways.  Once is for |SetoidFamily|
where the underlying |Setoid|s are different -- yields a kind of disjoint
union.  Is commutative.  But doesn't help with elem-of (xs ++ ys).

Another is for the same underlying |Setoid|.  The problem with this
one is that it does not want to be a congruence (code commented out).
It doesn't seem to be a coproduct either -- see the holes in the code.
ext for commute_2 is particularly telling!